+++ sdk/identity/azure-identity/samples/client_secret_credential.cpp
@@ -26,7 +26,7 @@ int main()
     Azure::Service::Client azureServiceClient("serviceUrl", clientSecretCredential);
 
     // Step 3: Start using the Azure Service Client.
-    azureServiceClient.DoSomething(Azure::Core::Context::ApplicationContext);
+    azureServiceClient.DoSomething();
 
     std::cout << "Success!" << std::endl;
   }