+++ sdk/core/azure-core/CHANGELOG.md
@@ -6,11 +6,17 @@
 
 ### Breaking Changes
 
+- Deprecated the `Azure::Core::Context::ApplicationContext` object. 
+  - If customer code is using `Azure::Core::Context::ApplicationContext`, the customer should instead create their own root context object which is used
+  wherever the customer would have previously used `Azure::Core::Context::ApplicationContext`.
+
 ### Bugs Fixed
 
 ### Other Changes
 
 - Updated JSON library to 3.11.3.
+- `Azure::Core::Context::ApplicationContext` has been deprecated. 
+- A new constructor was added to Azure::Core::Context to create a new context with a deadline directly.
 
 ## 1.13.0-beta.1 (2024-06-06)
 
@@ -22,6 +28,7 @@
 
 - [[#5622]](https://github.com/Azure/azure-sdk-for-cpp/pull/5622) Documentation fix for building the SDK with specific OpenSSL version.  (A community contribution, courtesy of _[ByteYue](https://github.com/ByteYue)_)
 - [[#5515]](https://github.com/Azure/azure-sdk-for-cpp/issues/5515) Add a `ShouldRetry` virtual method to the retry policy to enable customization of service-specific retry logic.
+   
 
 ### Acknowledgments